pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.84",
    "bs4>=0.0.2",
    "cohere>=5.21.1",
    "fastapi>=0.135.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.26",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.6",
    "langgraph-checkpoint>=4.0.1",
    "opensearch-py>=3.1.0",
    "python-dotenv>=1.2.2",
    "requests-aws4auth>=1.3.1",
    "uvicorn>=0.44.0",
]
